+++ CHANGELOG.md
@@ -26,6 +26,7 @@ CHANGELOG
 - Avoid failing on DescribeCluster when cluster configuration is not available.
 - Fix `UpdateParallelClusterLambdaRole` in the ParallelCluster API to upload logs to CloudWatch.
 - Fix Cinc not using the local CA certificates bundle when installing packages before any cookbooks are executed.
+- Fix a hang in upgrading ubuntu via `pcluster build-image` when `Build:UpdateOsPackages:Enabled:true` is set.
 
 3.2.0
 ------